+++ source/Plugins/Plugin_IniFile.pas
@@ -6,13 +6,13 @@ uses Classes, Plugin, uPSRuntime, uPSCompiler, PluginsMapFactory, API_IniFile,
   uPSI_API_IniFile, API_Output, SysUtils;
 
 type
-  tPlugin_RegIni = class(Tplugin)
+  tPlugin_RegIni = class(TInternalplugin)
   private
   protected
     foAPI_IniFile: TAPI_IniFile;
   public
     constructor Create(aAPI_Output: tAPI_Output;
-      var aImp: TPSRuntimeClassImporter); override;
+      var aImp: TPSRuntimeClassImporter;aPluginName: string); override;
     destructor Destroy; override;
 
     function CustomOnUses(var aCompiler: TPSPascalCompiler): Boolean; override;
@@ -24,7 +24,7 @@ type
 implementation
 
 constructor tPlugin_RegIni.Create(aAPI_Output: tAPI_Output;
-  var aImp: TPSRuntimeClassImporter);
+  var aImp: TPSRuntimeClassImporter;aPluginName: string);
 begin
   Inherited;